+++ alta_env/metadata.rb
@@ -4,5 +4,5 @@ maintainer_email 'seijiro.moriyama@anagix.com'
 license          'All rights reserved'
 description      'Installs/Configures alta_env'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.2.6'
+version          '0.2.7'
 depends 'windows'