+++ klayout_source_build/metadata.rb
@@ -1,4 +1,4 @@
-name 'klayout'
+name 'klayout_source_build'
 maintainer 'The Authors'
 maintainer_email 'you@example.com'
 license 'All Rights Reserved'